Experimenting with TGUi - added a sample button

+++ src/slot.cpp
@@ -1,6 +1,6 @@
 #include "slot.hpp"
 
-Slot::Slot(int x, int y, const sf::Color& c=sf::Color::White) {
+Slot::Slot(float x, float y, const sf::Color& c=sf::Color::White) {
     setPosition(sf::Vector2f(x, y));
     setFillColor(c);
     setOutlineColor(sf::Color::Black);
@@ -13,7 +13,7 @@ void Slot::addNeighbor(Slot& slot) {
     mNeighbors.push_back(&slot);
 }
 
-bool Slot::clicked(int mouseX, int mouseY) {
+bool Slot::clicked(float mouseX, float mouseY) {
     const sf::Vector2f* pos = &getPosition();
     bool x = (pos->x - RADIUS <= mouseX) && (mouseX <= pos->x + RADIUS);
     return x && (pos->y - RADIUS <= mouseY) && (mouseY <= pos->y + RADIUS);
